+++ src/Exceptions/InvalidArgumentException.php
@@ -1,0 +1,17 @@
+<?php
+/*
+ * This file is part of the Bushido\Foundation package.
+ *
+ * (c) Wojciech Nowicki <wnowicki@me.com>
+ *
+ * For the full copyright and license information, please view the LICENSE.md
+ * file that was distributed with this source code.
+ */
+
+namespace Bushido\Foundation\Exceptions;
+
+use Bushido\Foundation\Exception;
+
+class InvalidArgumentException extends Exception
+{
+}