+++ CHANGELOG.md
@@ -3,6 +3,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+ 
+## [Unreleased]
+### Added
+- `InvalidArgumentException` exception
+- `ArrayableHelper` class
+- `ChangeCase` class
+- `Entity` interface
+- `FlexEntity` class
 
 ## 0.0.1 - 2019-11-20
 ### Added